refactor(test): remove GetEntryGateway tests and implement the same logic in the EntryRepository
