+++ frontend/tests/Unit/Infrastructure/Repositories/Entries/GetById/AC02_Non2xxResponse.test.ts
@@ -1,0 +1,54 @@
+import {describe, it, expect, beforeEach, afterEach} from 'vitest';
+import {
+    createRepository,
+    mockJsonOnce,
+    type RepositoryTestCtx
+} from '@tests/Unit/Infrastructure/Repositories/Entries/BaseEntriesRepositoryTest';
+import {ac02Non2xxResponse as makeRequest} from '@tests/helpers/http/requests/entries/GetEntryRequestFactory';
+import {
+    makeBadRequest,
+    makeInternalError
+} from '@tests/helpers/http/responses/common/Non2xxResponseFactory';
+
+/**
+ * UC-3: Get Entry (Repository)
+ * Verifies EntryRepository.findById rejects on generic non-2xx HTTP responses (400/500).
+ * Mechanics: build request via factory, enqueue JSON with 400/500, expect rejection.
+ * @covers EntryRepository
+ */
+describe('AC02 — EntryRepository.findById throws on non-2xx response (generic)', () => {
+    let ctx: RepositoryTestCtx;
+
+    beforeEach(() => {
+        ctx = createRepository();
+    });
+    afterEach(() => {
+        ctx.cleanup();
+    });
+
+    it('throws when API responds with 400 Bad Request', async () => {
+        // Arrange
+        const request = makeRequest();
+        const response = makeBadRequest();
+        mockJsonOnce(ctx.fetchMock, 400, response);
+
+        // Act
+        const fn = ctx.repo.findById(request.id);
+
+        // Assert
+        await expect(fn).rejects.toThrowError(/400|bad request/i);
+    });
+
+    it('throws when API responds with 500 Internal Server Error', async () => {
+        // Arrange
+        const request = makeRequest();
+        const response = makeInternalError();
+        mockJsonOnce(ctx.fetchMock, 500, response);
+
+        // Act
+        const fn = ctx.repo.findById(request.id);
+
+        // Assert
+        await expect(fn).rejects.toThrowError(/500|internal/i);
+    });
+});
